refactor: extract useIsLockScreen hook with platform-specific files
- useIsLockScreen.ts: listens to LockApp/UnlockApp events, returns
  reactive lock screen state
- useIsLockScreen.native.ts: noop, always returns false (no trapFocus
  on native, no need for event listeners)
- Dialog uses the hook instead of inline event listeners
