+++ packages/components/src/hooks/useIsLockScreen.native.ts
@@ -1,0 +1,3 @@
+export function useIsLockScreen() {
+  return false;
+}
